Fixed error with singleplayer questions.

+++ WebQuizGame/Classes/AnswerRecord.cs
@@ -4,5 +4,5 @@ public class AnswerRecord
 {
     public Question Question { get; set; }
     public bool IsCorrect { get; set; }
-    public int TimeTaken { get; set; }
+    public double TimeTaken { get; set; }
 }
